# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "pdf-lib-workers",
    "version": "0.1.0",
...
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    }
  },
  "servers": [
    {
      "url": "https://pdf-lib-workers.{subdomain}.workers.dev",
      "description": "Your deployment; a custom domain works the same way. GET /openapi.json on a deployment lists its real address here.",
      "variables": {
        "subdomain": {
          "default": "your-subdomain",
          "description": "Your workers.dev subdomain."
        }
      }
    }
  ],
  "security": [
    {
      "bearerAuth": []
    },
    {
      "apiKeyHeader": []
    }
  ],
...
  "paths": {
...
    "/openapi.json": {
      "get": {
        "operationId": "openapi",
        "tags": [
          "Utilities"
        ],
        "summary": "This document",
        "security": [
          {}
        ],
        "responses": {
          "200": {
            "description": "The OpenAPI document.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/files/{key}": {
      "get": {
        "operationId": "downloadFile",
        "tags": [
          "Files"
        ],
        "summary": "Download a stored result",
        "description": "Use the signed `url` from a response (no API key needed until it expires), or send the API key with just the path.",
        "security": [
          {},
          {
            "bearerAuth": []
          },
          {
            "apiKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "description": "The R2 key, slashes included (e.g. \"outputs/abc.pdf\"). \"%2F\" is accepted for \"/\".",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "expires",
            "in": "query",
            "required": false,
            "description": "Signed links: expiry, as Unix seconds.",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "sig",
            "in": "query",
            "required": false,
            "description": "Signed links: signature.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "download",
            "in": "query",
            "required": false,
            "allowEmptyValue": true,
            "description": "Present: ask the browser to save the file.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Range",
            "in": "header",
            "required": false,
            "description": "Part of the file, e.g. \"bytes=0-1023\".",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The file.",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "206": {
            "description": "Part of the file (Range request)."
          },
          "304": {
            "description": "Not modified (If-None-Match / If-Modified-Since)."
          },
          "401": {
            "description": "No signature and no valid API key.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Signed link invalid or expired.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "No file at this key.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "head": {
        "operationId": "checkFile",
        "tags": [
          "Files"
        ],
        "summary": "Check a stored result without downloading it",
        "description": "Use the signed `url` from a response (no API key needed until it expires), or send the API key with just the path.",
        "security": [
          {},
          {
            "bearerAuth": []
          },
          {
            "apiKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "description": "The R2 key, slashes included (e.g. \"outputs/abc.pdf\"). \"%2F\" is accepted for \"/\".",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "expires",
            "in": "query",
            "required": false,
            "description": "Signed links: expiry, as Unix seconds.",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "sig",
            "in": "query",
            "required": false,
            "description": "Signed links: signature.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "download",
            "in": "query",
            "required": false,
            "allowEmptyValue": true,
            "description": "Present: ask the browser to save the file.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Range",
            "in": "header",
            "required": false,
            "description": "Part of the file, e.g. \"bytes=0-1023\".",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The file.",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "206": {
            "description": "Part of the file (Range request)."
          },
          "304": {
            "description": "Not modified (If-None-Match / If-Modified-Since)."
          },
          "401": {
            "description": "No signature and no valid API key.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Signed link invalid or expired.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "No file at this key.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/pdf/info": {
      "post": {
        "operationId": "getInfo",
        "tags": [
          "Inspect"
        ],
        "summary": "Pages, boxes, metadata, form fields, layers, viewer preferences, attachments",
        "parameters": [
          {
            "name": "options",
            "in": "query",
            "required": false,
            "description": "With a raw PDF body: the rest of the request as JSON, URL-encoded.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InfoRequest"
              }
            },
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "options": {
                    "type": "string",
                    "description": "The InfoRequest as JSON. Files in this request are used with { \"upload\": \"<field or file name>\" }."
                  }
                },
                "additionalProperties": {
                  "type": "string",
                  "format": "binary",
                  "description": "A file. Field names may repeat; repeated ones are addressed as name[0], name[1], …"
                }
              },
              "encoding": {
                "options": {
                  "contentType": "application/json"
                }
              }
            },
            "application/pdf": {
              "schema": {
                "type": "string",
                "format": "binary",
                "description": "The PDF itself, used as the source. Put the rest of the request in ?options=."
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "An InfoResponse, or a LockedInfoResponse for an encrypted PDF sent without its password.",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/InfoResponse"
                    },
                    {
                      "$ref": "#/components/schemas/LockedInfoResponse"
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid request: bad JSON or fields (see details), a page out of range, an unknown form field, a font missing characters.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or wrong API key.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "No file at an R2 key used as a source.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "A URL source is larger than the MAX_FETCH_BYTES setting.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Not a PDF, a damaged PDF, a wrong or missing password, or an operation the PDF cannot support.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "API_KEY is not set, or an unexpected error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "A URL source returned an error or could not be reached.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "504": {
            "description": "A URL source timed out.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "PdfSource": {
        "description": "A PDF: a PdfSourceObject or a shortcut string.",
        "anyOf": [
          {
            "type": "string",
            "minLength": 1,
            "description": "Shortcut for a source object: a string starting with \"http://\" or \"https://\" is { url }, one starting with \"data:\" is { base64 }, anything else is { key }."
          },
          {
            "$ref": "#/components/schemas/PdfSourceObject"
          }
        ]
      },
...
      "InfoRequest": {
        "type": "object",
        "properties": {
          "source": {
            "$ref": "#/components/schemas/PdfSource"
          }
        },
        "required": [
          "source"
        ],
        "description": "A locked PDF sent without its password returns a LockedInfoResponse."
      },
...
      "LockedInfoResponse": {
        "type": "object",
        "properties": {
          "pageCount": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "encrypted": {
            "type": "boolean",
            "const": true
          },
          "needsPassword": {
            "type": "boolean",
            "const": true
          },
          "pages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PageInfo"
            }
          }
        },
        "required": [
          "pageCount",
          "encrypted",
          "needsPassword",
          "pages"
        ],
        "description": "An encrypted PDF sent without its password: only its structure is readable."
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "What went wrong. Names the failing input, e.g. sources[1] or operations[2] (removePages)."
          },
          "details": {
            "description": "For 400 validation errors: each invalid field.",
            "anyOf": [
              {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "path": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "path",
                    "message"
                  ]
                }
              },
              {
                "type": "string"
              }
            ]
          }
        },
        "required": [
          "error"
        ]
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Authorization: Bearer <API_KEY>"
      },
      "apiKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  }
}